+++ assets/javascripts/kitten/components/form/radio-button/radio-button.stories.js
@@ -25,6 +25,11 @@ export const Default = () => (
         large={boolean('Large', false)}
         largeContent={boolean('LargeContent', false)}
         text={text('Text', 'RadioButton Label')}
+        size={select(
+          'Size',
+          { regular: 'regular', big: 'big' },
+          'regular',
+        )}
         variant={select(
           'Variant',
           { andromeda: 'andromeda', orion: 'orion' },
@@ -49,6 +54,11 @@ export const Default = () => (
           { andromeda: 'andromeda', orion: 'orion' },
           'andromeda',
         )}
+        size={select(
+          'Size',
+          { regular: 'regular', big: 'big' },
+          'regular',
+        )}
         design={select('Design', { disc: 'disc', check: 'check' }, 'radio')}
       >
         {text('Content', '')}
@@ -67,6 +77,11 @@ export const Default = () => (
           'Variant',
           { andromeda: 'andromeda', orion: 'orion' },
           'andromeda',
+        )}
+        size={select(
+          'Size',
+          { regular: 'regular', big: 'big' },
+          'regular',
         )}
         design={select('Design', { disc: 'disc', check: 'check' }, 'radio')}
       >