+++ assets/javascripts/kitten/components/form/radio-button/test.js
@@ -1,61 +1,31 @@
 import React from 'react'
-import { RadioButton } from '../../../components/form/radio-button'
 import renderer from 'react-test-renderer'
+import 'jest-styled-components'
+import { RadioButton } from './index'
 
-describe('<RadioButton />', () => {
-  describe('By default', () => {
-    const component = shallow(<RadioButton id="test-id" text="Label Text" />)
-    const input = component.find('input')
-
-    it('renders an input.k-Form-RadioButton__input">', () => {
-      expect(input.is('input')).toBe(true)
-      expect(input.hasClass('k-Form-RadioButton__input')).toBe(true)
-    })
-  })
-
-  describe('large prop', () => {
-    const component = shallow(
-      <RadioButton id="test-id" text="Label Text" large />,
-    )
-    const radioButton = component.find('.k-Form-RadioButton')
-
-    it('passes the right props to the component', () => {
-      expect(radioButton.hasClass('k-Form-RadioButton--largeLabel')).toBe(true)
-    })
-  })
-
-  describe('largeContent prop', () => {
-    const component = shallow(
-      <RadioButton id="test-id" text="Label Text" largeContent>
-        Hello
-      </RadioButton>,
-    )
-    const radioButton = component.find('.k-Form-RadioButton')
-
-    it('passes the right props to the component', () => {
-      expect(radioButton.hasClass('k-Form-RadioButton--largeContent')).toBe(
-        true,
-      )
-    })
-  })
-
-  describe('error prop', () => {
-    const component = shallow(
-      <RadioButton id="test-id" text="Label Text" error />,
-    )
-    const input = component.find('.k-Form-RadioButton')
-
-    it('passes the right props to the `input`component', () => {
-      expect(input.hasClass('k-Form-RadioButton--error')).toBe(true)
-    })
-  })
+describe('<RadiButton />', () => {
+  let component
 
   describe('default', () => {
-    let component
     beforeEach(() => {
       component = renderer
-        .create(<RadioButton id="test-id" text="Text" />)
-        .toJSON()
+      .create(
+        <RadioButton
+          id="test_1"
+          name="test"
+          error={false}
+          disabled={false}
+          large={false}
+          largeContent={false}
+          text="RadioButton Label"
+          size="regular"
+          variant="andromeda"
+          design="disc"
+        >
+          Children
+        </RadioButton>
+      )
+      .toJSON()
     })
 
     it('matches with snapshot', () => {